requirements.txt
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
tensorflow>=2.14.0
nltk>=3.8.1
langchain>=0.1.0
langchain-community>=0.0.20
anthropic>=0.18.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
regex>=2023.10.3
textblob>=0.17.1
streamlit-chat>=0.1.1
requests>=2.31.0
joblib>=1.3.2
tqdm>=4.66.1
setuptools>=68.0.0